transferred logic from Activities into a new Activity component
